app/requirements.txt
flask
snowflake-connector-python
python-dotenv
